# Messenger/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API для мессенджера",
    "version": "v1"
  },
  "paths": {
    "/api/auth": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Получить токен аунтетификации пользователя.",
        "description": "Пример запроса:\r\n    POST /Auth\r\n    {\r\n        \"username\": \"user\",\r\n        \"password\": \"password\"\r\n    }",
        "requestBody": {
          "description": "Данные пользователя",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserAuthInfo"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserAuthInfo"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserAuthInfo"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessTokenJSON"
                }
              }
            }
          },
          "400": {
            "description": "Неправильное имя пользователя или пароль"
          }
        }
      }
    },
    "/api/file": {
      "get": {
        "tags": [
          "File"
        ],
        "summary": "Получить файл с заданным индетификатором.",
        "parameters": [
          {
            "name": "fileHandle",
            "in": "query",
            "description": "Индетификатор файла",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Файл"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Файл не найден"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "post": {
        "tags": [
          "File"
        ],
        "summary": "Загрузить файл для сообщения. Файл должен быть загружен с помощью form-data.",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "file": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Уникальный id файла который используется для последуюзего обращения к нему",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FileHandleJSON"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileHandleJSON"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileHandleJSON"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/images": {
      "post": {
        "tags": [
          "Images"
        ],
        "summary": "Загрузить аватарку пользователя. Пользователь определяется JWT токеном при аунтетификации.\r\nФайл изображения должен быть формата .png или .jpg и загружен с помощью form-data",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "file": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "400": {
            "description": "Загруженный файл не является изображением"
          },
          "409": {
            "description": "Ошибка при загрузке изображения"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/images/{userID}": {
      "get": {
        "tags": [
          "Images"
        ],
        "summary": "Получить аватарку пользователя. Требует аунтетификации.",
        "parameters": [
          {
            "name": "userID",
            "in": "path",
            "description": "Идентификатор пользователя для которого требуется получить аватрку.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Указанный пользователь не нейден"
          },
          "400": {
            "description": "Не найден аватар пользователя"
          },
          "409": {
            "description": "Ошибка загрузки файла"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/messenger/history": {
      "get": {
        "tags": [
          "Message"
        ],
        "summary": "Получение истории сообщений для конкретного пользователя. Пользователь определяется по JWT токену.",
        "responses": {
          "200": {
            "description": "История сообщений",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Message"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Ошибка авторизации"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/messenger/history/{userID}": {
      "get": {
        "tags": [
          "Message"
        ],
        "summary": "Получить историю сообщений для двух конкретных пользователей. Первый пользователь по JWT.",
        "parameters": [
          {
            "name": "userID",
            "in": "path",
            "description": "Второй пользователь",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "История сообщений",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Message"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Ошибка авторизации"
          },
          "404": {
            "description": "Пользователь с заданным userId не был найден"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/messenger/connect": {
      "get": {
        "tags": [
          "Message"
        ],
        "summary": "Endpoint для подключения к мессенджеру по websocket.",
        "parameters": [
          {
            "name": "accessToken",
            "in": "query",
            "description": "JWT-токен доступа",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Попытка подключения не по протоколу websocket"
          },
          "404": {
            "description": "Указанный пользователь не найден"
          }
        }
      }
    },
    "/api": {
      "get": {
        "tags": [
          "Messenger"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/user/all": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Получения списка всех пользователей. Требует аунтетификации.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClientUserInfo"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Ошибка аунтетификации."
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/user": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Получение информации о конкретном пользователе. Требует аунтетификации.",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Идентификатор пользователя",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientUserInfo"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Пользователь не найден"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "patch": {
        "tags": [
          "User"
        ],
        "summary": "Изменить информацию пользователя. Пользователь определяется по jwt-токену.",
        "requestBody": {
          "description": "Информация для изменения",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserPatchInfo"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserPatchInfo"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserPatchInfo"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientUserInfo"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Пользователь с данным ID не найден"
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/user/register": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Регистрация пользователя",
        "requestBody": {
          "description": "Информация о пользователе, которого необходимо зарегистрировать",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegisterInfo"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegisterInfo"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegisterInfo"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUser"
                }
              }
            }
          },
          "409": {
            "description": "Пользователь с данным именем уже существует"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AccessTokenJSON": {
        "type": "object",
        "properties": {
          "accesToken": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AuthUser": {
        "required": [
          "nickname"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Индетификатор пользователя",
            "format": "uuid"
          },
          "nickname": {
            "type": "string"
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Информация о пользователе"
      },
      "ClientUserInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nickname": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "FileHandleJSON": {
        "type": "object",
        "properties": {
          "fileHandle": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
      "Message": {
        "type": "object",
        "properties": {
          "userFrom": {
            "type": "string",
            "description": "Пользователь, отправивший сообщение.",
            "format": "uuid"
          },
          "userTo": {
            "type": "string",
            "description": "Пользователь, кому назначено сообщение.",
            "format": "uuid"
          },
          "messageType": {
            "$ref": "#/components/schemas/MessageType"
          },
          "content": {
            "type": "string",
            "description": "Текстовое содержание сообщения.",
            "nullable": true
          },
          "fileID": {
            "type": "string",
            "description": "Файл, содержащийся в сообщении (при наличии)",
            "format": "uuid",
            "nullable": true
          },
          "dateSent": {
            "type": "string",
            "description": "Дата отправки сообщения.",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "Сообщение"
      },
      "MessageType": {
        "enum": [
          "Text",
          "File"
        ],
        "type": "string",
        "description": "Тип контента сообщения"
      },
      "UserAuthInfo": {
        "required": [
          "password",
          "username"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "Имя пользователя"
          },
          "password": {
            "type": "string",
            "description": "Пароль пользователя"
          }
        },
        "additionalProperties": false,
        "description": "Информация о пользователе для авторизации"
      },
      "UserPatchInfo": {
        "type": "object",
        "properties": {
          "nickname": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UserRegisterInfo": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "nickname": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT-токен для авторизации",
        "scheme": "Bearer"
      }
    }
  }
}